# depression/docs/openapi_v1.yaml
openapi: 3.0.3
info:
  title: 抑郁症辅助干预系统 OpenAPI 接口草稿（V1）
  version: 1.0.0-draft
  description: |
    本文档是抑郁症辅助干预系统 V1 阶段的 OpenAPI（接口描述标准）草稿。

    当前草稿基于以下三个设计文档整理而来：
    1. system_modular_api_design_v1_20260401.md
    2. system_module_master_table_v1_20260401.md
    3. system_api_response_standard_v1_20260401.md

    当前目标：
    1. 固定 11 个一级模块和 12 个二级能力模块的主接口
    2. 固定统一返回结构 code / message / data / request_id / timestamp
    3. 为后续 Flask（Web 框架）或 APIFlask（基于 Flask 的 API 框架）实现提供代码依据
    4. 为组员联调、Mock（模拟接口）和后续服务化改造提供统一契约

    当前范围：
    1. 每个二级能力模块先提供 1 个主业务接口
    2. 每个二级能力模块必须提供 1 个 health（健康检查）接口
    3. 每个二级能力模块必须提供 1 个 meta（元信息）接口
servers:
  - url: http://127.0.0.1:5000
    description: 本地 Flask（Web 框架）开发环境
  - url: http://127.0.0.1:8000
    description: 预留服务化或独立 API（接口）环境
...
paths:
...
  /api/v1/risk-assessment/risk-stratification/score:
    post:
      tags:
        - risk-assessment
      summary: 初步风险分层与安全建议能力主接口
      operationId: riskAssessmentRiskStratificationScore
      description: |
        `risk-stratification（初步风险分层与安全建议能力）` 的主接口。
        用于聚合建档信息、量表结果和作答过程的多模态摘要结果，
        输出初步风险等级、阈值命中情况、就医建议和是否允许继续体验系统调控。
        该接口不直接负责最终的个性化调控内容选择。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskStratificationScoreRequest'
            examples:
              standard_case:
                summary: 常规评估请求示例
                value:
                  user_id: user-1001
                  session_id: sess-risk-001
                  user_profile:
                    age_group: 18-25
                    gender: female
                    education_stage: undergraduate
                    living_status: dormitory
                  persona_preference:
                    persona_id: warm-doctor-01
                    style: warm
                    voice_preference: female-soft
                  assessment_summary:
                    available: true
                    scale_type: PHQ-9
                    total_score: 16
                    severity: moderately_severe
                    confidence: 0.95
                    summary: PHQ-9 得分较高，近两周抑郁症状较明显
                    self_harm_item_positive: false
                    structured_flags:
                      - sleep_problem
                      - low_mood
                  text_summary:
                    available: true
                    score: 0.44
                    level: medium
                    confidence: 0.73
                    summary: 用户补充提到近期学习压力大、睡眠差
                    keywords:
                      - 压力大
                      - 睡眠差
                  speech_summary:
                    available: true
                    score: 0.62
                    level: medium
                    confidence: 0.78
                    label: sad
                    summary: 语音整体低沉，语速偏慢
                  emotion_summary:
                    available: true
                    score: 0.58
                    level: medium
                    confidence: 0.72
                    label: sad
                    summary: 表情整体偏消极
                  eeg_summary:
                    available: false
                    score: null
                    level: null
                    confidence: null
                    summary: null
                  decision_context:
                    source_page: assessment_result
                    threshold_version: v1
                    collected_modalities:
                      - assessment
                      - text
                      - speech
                      - emotion
      responses:
        '200':
          description: 风险分层成功返回 / Risk stratification success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskStratificationResponse'
              examples:
                continue_to_dialogue:
                  summary: 中风险，继续进入引导问诊
                  value:
                    code: 0
                    message: ok
                    data:
                      result:
                        risk_level: medium
                        risk_score: 57
                        severity_band: medium
                        referral_recommended: false
                        emergency_notice: false
                        allow_continue: true
                        recommended_next_step: continue_to_dialogue
                        assessment_confidence: 0.84
                        triggered_rules: []
                        evidence_summary:
                          - source: assessment
                            signal: phq9_score
                            detail: 量表结果提示持续情绪困扰
                          - source: speech
                            signal: low_arousal_voice
                            detail: 语音表现出一定低沉和疲惫倾向
                        contributions:
                          assessment: 0.74
                          text: 0.08
                          speech: 0.10
                          emotion: 0.08
                          eeg: 0.0
                        safety_advice:
                          headline: 建议继续完成引导问诊，以补充近期状态和需求信息
                          actions:
                            - 保持后续问诊流程
                            - 继续观察睡眠与情绪波动
                        threshold_snapshot:
                          referral_threshold: high
                          emergency_threshold: urgent
                    request_id: req-20260402-risk-001
                    timestamp: '2026-04-02T10:30:00+08:00'
                referral_recommended_but_continue:
                  summary: 高风险，优先建议就医但允许继续体验
                  value:
                    code: 1001
                    message: referral_recommended
                    data:
                      result:
                        risk_level: high
                        risk_score: 78
                        severity_band: high
                        referral_recommended: true
                        emergency_notice: false
                        allow_continue: true
                        recommended_next_step: continue_with_strong_warning
                        assessment_confidence: 0.88
                        triggered_rules:
                          - rule_id: phq9_high_score
                            reason: PHQ-9 总分进入较高风险区间
                          - rule_id: multimodal_negative_consensus
                            reason: 语音与表情结果共同提示明显低落
                        evidence_summary:
                          - source: assessment
                            signal: severe_scale_signal
                            detail: 量表结果已达到较高风险区间
                          - source: emotion
                            signal: negative_affect_cluster
                            detail: 多模态结果共同提示明显异常
                        contributions:
                          assessment: 0.70
                          text: 0.06
                          speech: 0.12
                          emotion: 0.12
                          eeg: 0.0
                        safety_advice:
                          headline: 建议优先寻求专业帮助，系统调控仅作为辅助支持
                          actions:
                            - 优先联系心理老师、医生或可信赖照护者
                            - 如愿意，仍可继续完成引导问诊和后续调控体验
                        threshold_snapshot:
                          referral_threshold: high
                          emergency_threshold: urgent
                    request_id: req-20260402-risk-002
                    timestamp: '2026-04-02T10:32:00+08:00'
        '400':
          description: 风险评估请求参数错误 / Risk assessment bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_assessment_summary:
                  summary: 缺少量表摘要
                  value:
                    code: 4001
                    message: 'missing required field: assessment_summary'
                    data:
                      field: assessment_summary
                      module: risk-assessment
                      capability: risk-stratification
                    request_id: req-20260402-risk-003
                    timestamp: '2026-04-02T10:35:00+08:00'
                invalid_assessment_payload:
                  summary: 量表摘要字段不完整
                  value:
                    code: 4002
                    message: 'invalid field format: assessment_summary.total_score is required when available=true'
                    data:
                      field: assessment_summary.total_score
                      module: risk-assessment
                      capability: risk-stratification
                    request_id: req-20260402-risk-004
                    timestamp: '2026-04-02T10:36:00+08:00'
        '401':
          description: 风险评估接口鉴权失败 / Risk assessment unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 4003
                message: auth failed
                data:
                  module: risk-assessment
                  capability: risk-stratification
                  reason: login required before risk assessment
                request_id: req-20260402-risk-005
                timestamp: '2026-04-02T10:37:00+08:00'
        '500':
          description: 风险评估内部计算异常 / Risk assessment internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 5001
                message: risk stratification engine failed
                data:
                  module: risk-assessment
                  capability: risk-stratification
                  stage: rule_evaluation
                request_id: req-20260402-risk-006
                timestamp: '2026-04-02T10:38:00+08:00'
        '503':
          description: 风险评估服务未就绪 / Risk assessment unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                threshold_not_ready:
                  summary: 阈值配置未加载
                  value:
                    code: 7001
                    message: threshold config not loaded
                    data:
                      module: risk-assessment
                      capability: risk-stratification
                      required_resource: threshold_config_v1
                    request_id: req-20260402-risk-007
                    timestamp: '2026-04-02T10:39:00+08:00'
                upstream_unavailable:
                  summary: 上游依赖不可用
                  value:
                    code: 6001
                    message: upstream service unavailable
                    data:
                      module: risk-assessment
                      capability: risk-stratification
                      upstream: profile-store
                    request_id: req-20260402-risk-008
                    timestamp: '2026-04-02T10:40:00+08:00'
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - data
        - request_id
        - timestamp
      properties:
        code:
          type: integer
          description: 非 0 业务错误码 / Non-zero business error code
          example: 4001
        message:
          type: string
          description: 错误描述 / Error message
          example: missing required field
        data:
          type: object
          nullable: true
          description: 失败时通常为 null
          additionalProperties: true
          example: null
        request_id:
          type: string
          description: 请求唯一标识 / Request identifier
          example: req-20260401-004
        timestamp:
          type: string
          format: date-time
          description: 响应时间 / Response timestamp
          example: '2026-04-01T15:33:00+08:00'
...
    RiskAssessmentUserProfile:
      type: object
      description: 风险评估阶段使用的建档基础信息与初始画像 / Intake profile for risk assessment
      properties:
        age_group:
          type: string
          description: 年龄段 / Age group
          example: 18-25
        gender:
          type: string
          description: 性别 / Gender
          example: female
        education_stage:
          type: string
          description: 教育阶段 / Education stage
          example: undergraduate
        living_status:
          type: string
          description: 居住状态 / Living status
          example: dormitory
        known_history:
          type: array
          description: 已知既往情况 / Known history
          items:
            type: string
    PersonaPreferenceSummary:
      type: object
      description: 虚拟医生与交互偏好摘要 / Persona preference summary
      properties:
        persona_id:
          type: string
          description: 虚拟医生编号 / Persona identifier
          example: warm-doctor-01
        style:
          type: string
          description: 交互风格 / Interaction style
          example: warm
        voice_preference:
          type: string
          description: 语音偏好 / Voice preference
          example: female-soft
    AssessmentRiskSummary:
      type: object
      required:
        - available
      properties:
        available:
          type: boolean
          description: 当前量表摘要是否可用 / Whether assessment summary is available
          example: true
        scale_type:
          type: string
          description: 量表类型 / Scale type
          example: PHQ-9
        total_score:
          type: number
          nullable: true
          description: 量表总分 / Total assessment score
          example: 16
        severity:
          type: string
          nullable: true
          description: 量表严重程度标签 / Severity label
          example: moderately_severe
        confidence:
          type: number
          format: float
          nullable: true
          description: 量表结果置信度 / Confidence
          example: 0.95
        summary:
          type: string
          nullable: true
          description: 量表摘要说明 / Assessment summary
          example: PHQ-9 得分较高，近两周抑郁症状较明显
        self_harm_item_positive:
          type: boolean
          nullable: true
          description: 自伤相关题项是否阳性 / Whether self-harm related item is positive
          example: false
        structured_flags:
          type: array
          description: 结构化风险标记 / Structured flags
          items:
            type: string
          example:
            - sleep_problem
            - low_mood
    ModalitySignalSummary:
      type: object
      required:
        - available
      properties:
        available:
          type: boolean
          description: 模态结果是否可用 / Whether modality result is available
          example: true
        score:
          type: number
          format: float
          nullable: true
          description: 归一化模态分数 / Normalized modality score
          example: 0.62
        level:
          type: string
          nullable: true
          description: 模态等级标签 / Modality level label
          example: medium
        confidence:
          type: number
          format: float
          nullable: true
          description: 模态置信度 / Modality confidence
          example: 0.78
        label:
          type: string
          nullable: true
          description: 主标签 / Dominant label
          example: sad
        summary:
          type: string
          nullable: true
          description: 模态摘要说明 / Modality summary
          example: 语音整体低沉，语速偏慢
        keywords:
          type: array
          description: 关键词列表 / Keyword list
          items:
            type: string
          example:
            - 压力大
            - 睡眠差
    RiskStratificationContext:
      type: object
      description: 初步风险分层上下文 / Risk stratification context
      properties:
        source_page:
          type: string
          description: 来源页面 / Source page
          example: assessment_result
        threshold_version:
          type: string
          description: 阈值版本 / Threshold version
          example: v1
        collected_modalities:
          type: array
          description: 当前已采集的模态 / Collected modalities
          items:
            type: string
            enum:
              - assessment
              - text
              - speech
              - emotion
              - eeg
          example:
            - assessment
            - text
            - speech
            - emotion
    RiskTriggeredRule:
      type: object
      required:
        - rule_id
        - reason
      properties:
        rule_id:
          type: string
          description: 规则编号 / Rule identifier
          example: phq9_high_score
        reason:
          type: string
          description: 规则触发原因 / Rule trigger reason
          example: PHQ-9 总分进入较高风险区间
    RiskEvidenceItem:
      type: object
      required:
        - source
        - signal
        - detail
      properties:
        source:
          type: string
          description: 证据来源模块 / Evidence source
          example: assessment
        signal:
          type: string
          description: 信号名称 / Signal name
          example: phq9_score
        detail:
          type: string
          description: 证据说明 / Evidence detail
          example: 量表结果提示持续情绪困扰
    RiskSafetyAdvice:
      type: object
      properties:
        headline:
          type: string
          description: 安全建议主提示 / Safety advice headline
          example: 建议继续完成引导问诊，以补充近期状态和需求信息
        actions:
          type: array
          description: 建议动作列表 / Suggested actions
          items:
            type: string
    RiskThresholdSnapshot:
      type: object
      properties:
        referral_threshold:
          type: string
          description: 就医建议阈值 / Referral threshold
          example: high
        emergency_threshold:
          type: string
          description: 紧急提示阈值 / Emergency threshold
          example: urgent
    RiskStratificationResult:
      type: object
      required:
        - risk_level
        - risk_score
        - severity_band
        - referral_recommended
        - emergency_notice
        - allow_continue
        - recommended_next_step
      properties:
        risk_level:
          type: string
          description: 风险等级 / Risk level
          enum:
            - low
            - medium
            - high
            - urgent
          example: medium
        risk_score:
          type: integer
          description: 风险分数，范围 0-100 / Risk score from 0 to 100
          example: 57
        severity_band:
          type: string
          description: 风险挡位 / Severity band
          enum:
            - low
            - medium
            - high
            - urgent
          example: medium
        referral_recommended:
          type: boolean
          description: 是否建议优先寻求专业帮助 / Whether referral is recommended
          example: false
        emergency_notice:
          type: boolean
          description: 是否触发紧急提示 / Whether emergency notice is triggered
          example: false
        allow_continue:
          type: boolean
          description: 是否允许继续后续体验 / Whether flow can continue
          example: true
        recommended_next_step:
          type: string
          description: 推荐下一步 / Recommended next step
          enum:
            - continue_to_dialogue
            - continue_with_strong_warning
            - seek_professional_help_first
            - seek_emergency_support_now
          example: continue_to_dialogue
        assessment_confidence:
          type: number
          format: float
          description: 评估置信度 / Assessment confidence
          example: 0.84
        triggered_rules:
          type: array
          description: 触发规则列表 / Triggered rules
          items:
            oneOf:
              - type: string
              - $ref: '#/components/schemas/RiskTriggeredRule'
        evidence_summary:
          type: array
          description: 证据摘要 / Evidence summary
          items:
            $ref: '#/components/schemas/RiskEvidenceItem'
        contributions:
          type: object
          description: 各模态贡献占比 / Contributions by modality
          additionalProperties:
            type: number
            format: float
          example:
            assessment: 0.74
            speech: 0.10
            emotion: 0.08
            eeg: 0.0
        safety_advice:
          $ref: '#/components/schemas/RiskSafetyAdvice'
        threshold_snapshot:
          $ref: '#/components/schemas/RiskThresholdSnapshot'
    RiskStratificationResponseData:
      type: object
      required:
        - result
      properties:
        result:
          $ref: '#/components/schemas/RiskStratificationResult'
    RiskStratificationResponse:
      type: object
      required:
        - code
        - message
        - data
        - request_id
        - timestamp
      properties:
        code:
          type: integer
          description: 业务状态码，成功为 0，高风险提醒可使用 1xxx / Business code
          example: 0
        message:
          type: string
          description: 人类可读信息 / Human-readable message
          example: ok
        data:
          $ref: '#/components/schemas/RiskStratificationResponseData'
        request_id:
          type: string
          description: 请求唯一标识 / Request identifier
          example: req-20260402-risk-001
        timestamp:
          type: string
          format: date-time
          description: 响应时间 / Response timestamp
          example: '2026-04-02T10:30:00+08:00'
...
    RiskStratificationScoreRequest:
      type: object
      required:
        - assessment_summary
      properties:
        user_id:
          type: string
          description: 用户 ID；如未显式传入，服务端可根据登录会话自动补全 / User identifier
          example: user-1001
        session_id:
          type: string
          description: 会话 ID / Session identifier
          example: sess-risk-001
        user_profile:
          $ref: '#/components/schemas/RiskAssessmentUserProfile'
        persona_preference:
          $ref: '#/components/schemas/PersonaPreferenceSummary'
        assessment_summary:
          $ref: '#/components/schemas/AssessmentRiskSummary'
        text_summary:
          $ref: '#/components/schemas/ModalitySignalSummary'
        speech_summary:
          $ref: '#/components/schemas/ModalitySignalSummary'
        emotion_summary:
          $ref: '#/components/schemas/ModalitySignalSummary'
        eeg_summary:
          $ref: '#/components/schemas/ModalitySignalSummary'
        decision_context:
          $ref: '#/components/schemas/RiskStratificationContext'
...